adds user movement and map tiles

+++ Room.cs
@@ -9,7 +9,7 @@ namespace DungeonGame
         int Y1 { get; set; }
         int Y2 { get; set; }
 
-        double[] GetCenter();
+        int[] GetCenter();
         bool IntersectsRoom(Room room);
     }
     public class Room : IRoom
@@ -28,11 +28,11 @@ namespace DungeonGame
         public int Y1 { get; set; }
         public int Y2 { get; set; }
 
-        public double[] GetCenter()
+        public int[] GetCenter()
         {
-            var x = Math.Floor(((double)X1 + (double)X2) / 2);
+            var x = Math.Floor(((double)(X1) + (double)X2) / 2);
             var y = Math.Floor(((double)Y1 + (double)Y2) / 2);
-            return new double[] { x, y };
+            return new int[] { (int)x, (int)y };
         }
 
         public bool IntersectsRoom(Room room)
